+++ bfabric_scripts/src/bfabric_scripts/feeder/file_attributes.py
@@ -1,10 +1,11 @@
 from __future__ import annotations
 
-import hashlib
 import warnings
 from pathlib import Path
 from pydantic import BaseModel
 import datetime
+
+from bfabric.transfer import md5_checksum
 
 
 class FileAttributes(BaseModel):
@@ -16,11 +17,9 @@ class FileAttributes(BaseModel):
     @classmethod
     def compute(cls, file: Path) -> FileAttributes:
         """Computes the file attributes from the file at the given path."""
-        with file.open("rb") as f:
-            hash = hashlib.file_digest(f, "md5")
         file_stat = file.stat()
         return FileAttributes(
-            md5_checksum=hash.hexdigest(),
+            md5_checksum=md5_checksum(file),
             file_date=datetime.datetime.fromtimestamp(file_stat.st_mtime),
             file_size=file_stat.st_size,
             filename=file.name,
@@ -40,8 +39,7 @@ def get_file_attributes(file_name_or_attributes: str) -> tuple[str, int, int, st
         file_stat = file_path.stat()
         file_size = file_stat.st_size
         file_unix_timestamp = int(file_stat.st_mtime)
-        with file_path.open("rb") as f:
-            md5_checksum = hashlib.file_digest(f, "md5").hexdigest()
-        return md5_checksum, file_unix_timestamp, file_size, filename
+        checksum = md5_checksum(file_path)
+        return checksum, file_unix_timestamp, file_size, filename
     else:
         raise ValueError("Invalid input line format")